skills/imagegen/pyproject.toml
[project]
name = "imagegen"
version = "0.1.0"
description = "AI image generation CLI with provider-specific subcommands"
requires-python = ">=3.11"
dependencies = [
    "typer",
    "openai",
    "google-genai",
    "python-dotenv",
]

[project.scripts]
imagegen = "imagegen.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/imagegen"]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "pyright",
]

[tool.pytest.ini_options]
testpaths = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pyright]
pythonVersion = "3.11"
include = ["src"]
typeCheckingMode = "basic"
